Move ProcessingSession into the main crate.
